Cleanup fuzzing mocks
