+++ fuzzing/mock/os/os_exceptions.c
@@ -1,11 +1,8 @@
-/* Exception and NVM mocks.
- * os_sched_exit and os_lib_end longjmp back to the fuzz harness.
- */
+/* Exception mocks: os_sched_exit and os_lib_end longjmp back to the fuzz harness. */
 
 #include <setjmp.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <string.h>
 
 #include "exceptions.h"
 #include "os_task.h"
@@ -55,18 +52,4 @@ void __attribute__((noreturn)) os_sched_exit(bolos_task_status_t exit_code)
 void __attribute__((noreturn)) os_lib_end(void)
 {
     longjmp(fuzz_exit_jump_ctx.jmp_buf, 1);
-}
-
-void nvm_write(void *dst_adr, void *src_adr, unsigned int src_len)
-{
-    if (dst_adr == NULL || src_len == 0) {
-        return;
-    }
-
-    if (src_adr == NULL) {
-        memset(dst_adr, 0, src_len);
-    }
-    else {
-        memcpy(dst_adr, src_adr, src_len);
-    }
 }
